add ws alternative
